fix: val 사용하지 않던 문제 해결
