+++ diazol/src/utils/putEmojiNum.ts
@@ -3,12 +3,12 @@ import { db } from "@/utils/firebase";
 import { arrayUnion, doc, updateDoc } from "firebase/firestore";
 
 /** store data to firestore */
-export const putUserData = async (content: UserDataType) => {
+export const putUserData = async (userData: UserDataType) => {
   const userDBRef = doc(db, "survey", "data");
   const date = new Date();
   try {
     await updateDoc(userDBRef, {
-      data: arrayUnion({ date, content }),
+      data: arrayUnion({ date, userData }),
     });
   } catch (e) {
     console.log(e);
